Encode_SYS/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
python-multipart
PyJWT
pypdf>=4.0
cdp-sdk
cryptography
web3>=6.0
